Bug Fix: Failed to deserialize the date field from a number using the @JsonFormat annotation.
